backend/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
pyotp==2.9.0
qrcode==7.4.2
python-dotenv==1.0.0
mysqlclient==2.2.0
Werkzeug==3.0.1
Pillow==10.1.0
cryptography==41.0.7
scikit-learn==1.3.2
python-logstash==0.4.8
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
